Final databinding for bar graph data pending. All other dashboard work done.

+++ BlazorCrud.Shared/Models/Dashboard.cs
@@ -4,16 +4,16 @@ namespace BlazorCrud.Shared.Models
 {
     public class Dashboard
     {
-        public Dictionary<string, int> PatientsByState { get; set; }
-        public Dictionary<string, int> OrganizationsByType { get; set; }
-        public Dictionary<string, int> ClaimsByType { get; set; }
+        public Dictionary<string, double> PatientsByState { get; set; }
+        public Dictionary<string, double> OrganizationsByType { get; set; }
+        public Dictionary<string, double> ClaimsByType { get; set; }
         public Dictionary<string, Dictionary<string, int>> UpdatedEntitiesByDate { get; set; }
 
         public Dashboard()
         {
-            PatientsByState = new Dictionary<string, int>();
-            OrganizationsByType = new Dictionary<string, int>();
-            ClaimsByType = new Dictionary<string, int>();
+            PatientsByState = new Dictionary<string, double>();
+            OrganizationsByType = new Dictionary<string, double>();
+            ClaimsByType = new Dictionary<string, double>();
             UpdatedEntitiesByDate = new Dictionary<string, Dictionary<string, int>>();
         }
     }
